♻️(types): move types in const.ts to types.ts
